refactor(engine): code quality improvements — extract struct, flatten nesting, clippy fixes
adversarial_supremacy_synthesis.rs:
- Extract WorkloadHashInput struct to replace 8-argument compute_hash()
  function, improving readability and reducing parameter confusion risk

lib.rs:
- Remove redundant return in match arm (clippy lint)

ifc_artifacts.rs:
- Flatten nested if into combined condition with &&

object_model.rs:
- Reformat Proxy match arm for consistency

stdlib.rs:
- Add missing semicolon on return Err(...)
- Refactor surrogate pair parsing using let-else chains

metadata_substrate_governance.rs:
- Remove unnecessary 'as u64' cast (value already u64)

runtime_decision_core.rs:
- Minor formatting cleanup

+++ crates/franken-engine/src/metadata_substrate_governance.rs
@@ -533,7 +533,7 @@ impl GovernanceReceipt {
             append_str(&mut buf, &v.summary);
         }
         // Include portability score and target coverage.
-        append_u64(&mut buf, self.portability_score_millionths as u64);
+        append_u64(&mut buf, self.portability_score_millionths);
         for t in &self.targets_covered {
             append_str(&mut buf, &format!("{t:?}"));
         }
